Move LValueOccurrenceCount and VariableOccurrenceCount into ExprUtil class
